Placed templates in their own files, fixed issues with width & max width.
