+++ src/managers/message-component-manager.js
@@ -2,6 +2,8 @@ import { log } from "../utilities/log";
 import { v4 as uuidv4 } from 'uuid';
 import { embedMessage } from "./message-manager";
 import { resolveMessageProperties } from "./gist-properties-manager";
+import { embedHTMLTemplate } from "../templates/embed";
+import { messageHTMLTemplate } from "../templates/message";
 const delay = ms => new Promise(res => setTimeout(res, ms));
 
 export function isElementLoaded(elementId) {
@@ -121,129 +123,14 @@ function showMessage() {
 }
 
 function embed(url, message) {
-  const wideOverlayPositions = ["x-gist-bottom", "x-gist-bottom", "x-gist-floating-top", "x-gist-floating-bottom"];
   var messageProperties = resolveMessageProperties(message);
-  var maxWidthBreakpoint = 800;
-  if (messageProperties.messageWidth > maxWidthBreakpoint) {
-    maxWidthBreakpoint = messageProperties.messageWidth;
-  }
-
-  var messageWidth = messageProperties.messageWidth + "px";
-  if (wideOverlayPositions.includes(messageProperties.elementId) && !messageProperties.hasCustomWidth) {
-    messageWidth = "100%";
-  }
-
-  var template = `
-  <div id="gist-embed">
-    <style>
-      #x-gist-floating-top, #x-gist-floating-top-left, #x-gist-floating-top-right {
-        position: fixed;
-        top: 0px;
-        z-index: 1000000;
-      }
-      #x-gist-floating-bottom, #x-gist-floating-bottom-left, #x-gist-floating-bottom-right {
-        position: fixed;
-        bottom: 0px;
-        z-index: 1000000;
-      }
-      #x-gist-bottom, #x-gist-top, #x-gist-floating-top, #x-gist-floating-bottom {
-        left: 50%;
-        transform: translate(-50%, 0%);
-        width: '${messageWidth}';
-      }
-      #x-gist-floating-top-left, #x-gist-floating-top-right, #x-gist-floating-bottom-left, #x-gist-floating-bottom-right {
-        width: '${messageWidth}';
-      }
-      #x-gist-floating-top-right, #x-gist-floating-bottom-right {
-        right: 0px;
-      }
-      #gist-embed {
-        position: relative;
-        height: 100%;
-        width: 100%;
-      }
-      #gist-embed-container {
-        position: relative;
-        height: 100%;
-        width: 100%;
-      }
-      #gist-embed-container .gist-frame {
-        height: 100%;
-        width: 100%;
-        border: none;
-      }
-      @media (max-width: '${maxWidthBreakpoint}px') {
-        #x-gist-bottom, #x-gist-bottom, #x-gist-floating-top, #x-gist-floating-bottom, #x-gist-floating-top-left, #x-gist-floating-top-right, #x-gist-floating-bottom-left, #x-gist-floating-bottom-right {
-          width: 100%;
-        }
-      }
-    </style>
-    <div id="gist-embed-container">
-      <iframe class="gist-frame" src="${url}"></iframe>
-    </div>
-  </div>
-  `;
+  var template = embedHTMLTemplate(messageProperties, url);
   return template;
 }
 
 function component(url, message) {
   var messageProperties = resolveMessageProperties(message);
-  var maxWidthBreakpoint = 600;
-  if (messageProperties.messageWidth > maxWidthBreakpoint) {
-    maxWidthBreakpoint = messageProperties.messageWidth;    
-  }
-  var template = `
-    <div id="gist-embed-message">
-    <style>
-      #gist-overlay.background {
-        position: fixed;
-        z-index: 9999999998;
-        left: 0;
-        top: 0;
-        width: 100%;
-        height: 100%;
-        background-color: '${messageProperties.overlayColor}';
-        visibility: hidden;
-      }
-      #gist-overlay.background.visible {
-        visibility: visible;
-      }
-      #gist-overlay.background.is-blacked-out {
-        display: block;
-      }
-      #gist-message {
-        width: '${messageProperties.messageWidth}px';
-        position: absolute;
-        border: none;
-        opacity: 0;
-        transition: opacity 0.3s ease-in-out;
-        z-index: 9999999999;
-        left: 50%;
-        transform: translateX(-50%);
-      }
-      #gist-message.visible {
-        opacity: 1;
-      }
-      #gist-message.center {
-        transform: translate(-50%, -50%);
-        top: 50%;
-      }
-      #gist-message.bottom {
-        bottom: 0;
-      }
-      #gist-message.top {
-        top: 0;
-      }
-      @media (max-width: '${maxWidthBreakpoint}px') {
-        #gist-message {
-          width: 100%;
-        }
-      }
-    </style>
-    <div id="gist-overlay" class="background">
-      <iframe id="gist-message" class="message" src="${url}"></iframe>
-    </div>
-  </div>`;
+  var template = messageHTMLTemplate(messageProperties, url);
   return template;
 }
